Decided the best name for the findInclusive/detect would be search(), and aliased find() as searchInside().
Fixed search() to properly handle selectors that start with the element searched on, but are looking for children (i.e: “$('#baseball.game').search('.game > .player')”). Now it works, but is terribly inefficient.

Although XPath may be slow as a whole, I really think it kicks Sizzle's ass in situations like this.