Moved Node and Prop views to sub-directories “modifiable”, “observable”, or the fallback, “default”.
Created an ApplicationHelper, “render_subscoped”, that tries the given sub-directory first; if that fails, it falls back. It's a bit messy, but it works for now; however, it would be ideal to hook it into Rails a bit better, or at least make it soley a path-generation method, not a render wrapper method.